+++ genti-api/src/main/java/com/gt/genti/auth/api/AuthApi.java
@@ -14,8 +14,8 @@ import com.gt.genti.swagger.EnumResponse;
 import com.gt.genti.swagger.EnumResponses;
 import com.gt.genti.user.model.OauthPlatform;
 import com.gt.genti.user.model.UserRole;
-import com.gt.genti.user.service.social.AppleAuthTokenDto;
-import com.gt.genti.user.service.social.KakaoAccessTokenDto;
+import com.gt.genti.auth.dto.request.AppleAuthTokenDto;
+import com.gt.genti.auth.dto.request.KakaoAccessTokenDto;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -31,12 +31,7 @@ import jakarta.validation.constraints.NotNull;
 
 @Tag(name = "[AuthController] 인증 컨트롤러", description = "로그인을 처리한 후 토큰을 전달합니다.")
 public interface AuthApi {
-	@Operation(summary = "oauth 로그인 페이지 호출", description = "구글, 카카오 oauth로그인페이지 로 Redirect 됩니다. url은 HttpHeader에 포함되어있습니다.")
-	@ApiResponses(value = {
-		@ApiResponse(responseCode = "303", description = "리디렉션을 위한 응답 헤더를 포함합니다.", headers = {
-			@Header(name = "Location", description = "리디렉션 URL", schema = @Schema(type = "string"))
-		}, content = @Content(schema = @Schema(hidden = true)))
-	})
+	@Operation(summary = "oauth 로그인 페이지 호출", description = "카카오 oauth 로그인 페이지 로 Redirect, 이후 로그인 성공시 다시 admin page로 redirect됩니다.(Genti token은 cookie에)")
 	RedirectView login(
 		@Parameter(description = "호출할 Oauth platform 종류", example = "KAKAO", schema = @Schema(allowableValues = {
 			"KAKAO"}))
@@ -63,7 +58,6 @@ public interface AuthApi {
 		@RequestBody @Valid KakaoAccessTokenDto tokenDto
 	);
 
-
 	void kakaoRedirectLogin(
 		HttpServletResponse response,
 		@RequestParam(name = "code") String code);
@@ -75,17 +69,10 @@ public interface AuthApi {
 	ResponseEntity<ApiResult<TokenResponse>> getTestJwt(
 		@NotNull @RequestParam(name = "role", value = "role") UserRole role);
 
-	@Operation(summary = "oauth platform에서 로그인 후 받은 토큰을 전달하여 가입/로그인", description = "현재 애플, 카카오 지원합니다")
+	@Operation(summary = "Genti 토큰 refresh api", description = "Genti 토큰 refresh api")
 	@EnumResponses(value = {
 		@EnumResponse(ResponseCode.OK),
-		@EnumResponse(ResponseCode.AppleOauthClaimInvalid),
-		@EnumResponse(ResponseCode.AppleOauthIdTokenExpired),
-		@EnumResponse(ResponseCode.AppleOauthIdTokenIncorrect),
-		@EnumResponse(ResponseCode.AppleOauthIdTokenInvalid),
-		@EnumResponse(ResponseCode.AppleOauthJwtValueInvalid),
-		@EnumResponse(ResponseCode.AppleOauthPublicKeyInvalid),
 	})
-
 	ResponseEntity<ApiResult<TokenResponse>> reissue(
 		@RequestBody @Valid TokenRefreshRequestDto tokenRefreshRequestDto
 	);
