+++ genti-api/src/main/java/com/gt/genti/auth/dto/request/KakaoAccessTokenDto.java
@@ -1,4 +1,4 @@
-package com.gt.genti.user.service.social;
+package com.gt.genti.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
