--- genti-api/src/main/java/com/gt/genti/auth/dto/request/SocialLoginRequest.java
@@ -1,8 +1,0 @@
-package com.gt.genti.auth.dto.request;
-
-import com.gt.genti.user.model.OauthPlatform;
-
-public interface SocialLoginRequest {
-	OauthPlatform getOauthPlatform();
-	String getCode();
-}
