+++ genti-api/src/main/java/com/gt/genti/auth/dto/response/OauthJwtResponse.java
@@ -1,7 +1,6 @@
 package com.gt.genti.auth.dto.response;
 
 import com.gt.genti.constants.JWTConstants;
-import com.gt.genti.user.model.UserRole;
 
 public record OauthJwtResponse(
 	String accessToken,
@@ -9,6 +8,7 @@ public record OauthJwtResponse(
 	String userRoleString
 ) {
 	public static OauthJwtResponse of(String accessToken, String refreshToken, String userRoleString) {
-		return new OauthJwtResponse(JWTConstants.JWT_PREFIX + accessToken, JWTConstants.JWT_PREFIX + refreshToken, userRoleString);
+		return new OauthJwtResponse(JWTConstants.JWT_PREFIX + accessToken, JWTConstants.JWT_PREFIX + refreshToken,
+			userRoleString);
 	}
 }
