--- genti-api/src/main/java/com/gt/genti/user/service/social/AppleAuthTokenResponse.java
@@ -1,19 +1,0 @@
-package com.gt.genti.user.service.social;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Getter;
-
-@Getter
-public class AppleAuthTokenResponse {
-	@JsonProperty("access_token")
-	String accessToken;
-	@JsonProperty("token_type")
-	String tokenType;
-	@JsonProperty("expires_in")
-	Long expiresIn;
-	@JsonProperty("refresh_token")
-	String refreshToken;
-	@JsonProperty("id_token")
-	String idToken;
-}
